fix: automatically disable companion features if too many errors occurred
The bot now automatically disables bot companion features if too many errors occurred fetching the data. This change applies to the activity feed, kill feed and the displayPlayerGearLevel option for the status embed. Once the bot hits the maximum amount of errors, it will disable the corresponding feature. It is possible to re-enable the feature using the update-server command after fixing the issue. The maximum amount of errors can be configured via `bot.max-failed-api-attempts`. The default value is 0 (unlimited amount of errors).

Note: this commit heavily refactors the bots update logic and error handling.