fix: automatically disable companion features on error
* fix: only resolve characters when displayPlayerGearLevel is true
* fix: allow users to use #channel syntax for channel-id parameters
* fix: automatically disable companion features if too many errors occurred

The bot now automatically disables bot companion features if too many errors occurred fetching the data. This change applies to the activity feed, kill feed and the displayPlayerGearLevel option for the status embed. Once a server hits the maximum amount of errors, the bot will disable the corresponding bot companion feature. It is possible to re-enable the feature using the update-server command after fixing the issue. The maximum amount of errors can be configured via `bot.max-failed-api-attempts`. The default value is 0 (unlimited amount of errors).

Note: this commit also heavily refactors the bots update logic and error handling