feat: encrypt the bot database at rest (#130)
Configures MVStore to encrypt the database at rest. I initially expected Nitrite to do this when the database is secured with a username and password but this doesn't seem to be the case. The bot will attempt to migrate existing databases automatically on startup. However, **existing backups are not migrated automatically** but new ones will be stored encrypted as well.